requirements.txt
flask
flask-cors
python-dotenv
flask-pymongo
requests